+++ src/utils/arrayUtils.ts
@@ -10,4 +10,8 @@ export class ArrayUtils {
     }, []);
   }
 
+  public static createNumberArray(count: number): Array<number> {
+    return Array.from(Array(count).keys());
+  }
+
 }